+++ src/main/java/domain/name/Name.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.name;
 
 public class Name {
     private static final int MINIMUM_NAME_LENGTH = 1;
