+++ src/test/java/domain/name/NameTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.name;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
