+++ src/test/java/domain/name/NamesTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.name;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
